Renaming test files in typechecker test.
The typechecking context current wellformedness implementation
requires all program names to be unique, in other words, the current
simple list implementation cannot handle scoping. Therefore, the most
straightforward way at the moment of preserving this property is to
rename the test files in the typechecker test. In any case, this is
what is done in the non-testing code.