Fix rule âApp.
The previous version was plain wrong because it was ignore the typ
argument passed to the function and instead was using 'arrowExistVar'
which is not defined at all.

The rule should also test the condition:
  Context.containsTypeUnassigned context typ

With this fix the Test.TestData4 program stoppped typechecking. After
looking at the program, it's obvious it should have never typechecked
in the first place.