Handle variants without pattern as a special case.
The previous version that converted a variant without pattern to a
variant with the empty tuple () broke variants that have type
patterns, which I want to introduce soon.

By not doing this conversion and instead treating the variant0 as a
special case, we can remove a special case in genPatternPredicate that
passed empty tuple to type patterns without arguments. This means we
can reuse types and type predicates with this code (e.g., a pattern of
Apple or a constructor with Fruit just work with the same code).