Make _ binders unique.
This fixes a bug in that the expanded names of "_" were not
unique. This normally doesn't show up because these binders are used
as arguments to small functions in tests. However, if used with result
patterns then the tests fail with the "_" binder already defined.

This fixes the bug by always giving a new name to "_" binders.